+++ packages/moengage_flutter/moengage_flutter_ios/ios/moengage_flutter_ios/Sources/MoEngageFlutterConstants.swift
@@ -44,5 +44,6 @@ struct MoEngageFlutterConstants{
         static let kInAppClickedCustomAction    = "onInAppCustomAction"
         static let kInAppDismissed              = "onInAppDismiss"
         static let kInAppSelfHandled            = "onInAppSelfHandle"
+        static let kLogoutComplete              = "onLogoutComplete"
     }
 }